+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Future release]
+## [0.3.0]
 
 ### Added
 
@@ -26,6 +26,7 @@
 
 - Add string to all arb files.
 
-[0.1.0]: https://github.com/lsaudon/l10nization/releases/tag/release-0.1.0
-[0.1.1]: https://github.com/lsaudon/l10nization/releases/tag/release-0.1.1
+[0.3.0]: https://github.com/lsaudon/l10nization/releases/tag/release-0.3.0
 [0.2.0]: https://github.com/lsaudon/l10nization/releases/tag/release-0.2.0
+[0.1.1]: https://github.com/lsaudon/l10nization/releases/tag/release-0.1.1
+[0.1.0]: https://github.com/lsaudon/l10nization/releases/tag/release-0.1.0